+++ wechat/miniprogram/api/user.ts
@@ -1,14 +1,14 @@
 import { CODE } from "@constant/error";
-import { URL } from "@constant/user";
+import { PATH } from "@constant/user";
 import { HttpError } from "@models/error";
 import error from "@utils/error";
 import http from "@utils/http";
 import logger from "@utils/logger";
-import type { DetailResponse, UpdateRequest, UpdateResponse } from "types/user";
+import type { DetailResponse, EditRequest, EditResponse } from "types/user";
 
 const detail = async () => {
 	try {
-		return await http.post<DetailResponse>(URL.DETAIL);
+		return await http.post<DetailResponse>(PATH.DETAIL);
 	} catch (e: unknown) {
 		logger.error("查询用户详情失败", e);
 
@@ -16,9 +16,9 @@ const detail = async () => {
 	}
 };
 
-const update = async (updated: UpdateRequest) => {
+const update = async (updated: EditRequest) => {
 	try {
-		return await http.post<UpdateResponse>(URL.UPDATE, updated);
+		return await http.post<EditResponse>(PATH.EDIT, updated);
 	} catch (e: unknown) {
 		logger.error("更新用户信息失败", e);
 
@@ -26,4 +26,4 @@ const update = async (updated: UpdateRequest) => {
 	}
 };
 
-export default { detail, update };
+export default { detail, edit: update };