+++ wechat/miniprogram/constant/error.ts
@@ -2,17 +2,14 @@ const CODE = {
 	UNKNOWN: 9999,
 	// 微信错误
 	WEIXIN: 1,
-	WEIXIN_LOGIN: 2,
-	WEIXIN_QR_CODE: 3,
-	WEIXIN_STORAGE_SET: 4,
-	// 权限错误
-	LOGIN: 1000,
+	WEIXIN_QR_CODE: 2,
 	// http 调用错误
 	HTTP: 2000,
 	// http 内部参数错误
 	HTTP_PARAMS: 2001,
 	// http 业务接口 - token
-	HTTP_API_ACCESS_TOKEN_LOGIN: 2100,
+  HTTP_API_ACCESS_TOKEN_LOGIN: 2100,
+  HTTP_API_ACCESS_TOKEN_VALID: 2101,
 	// http 业务错误 - 用户接口
 	HTTP_API_USER_DETAIL: 2201,
 	HTTP_API_USER_UPDATE: 2202,
@@ -29,13 +26,11 @@ const CODE = {
 const MESSAGE = {
 	[CODE.UNKNOWN]: "未知错误",
 	[CODE.WEIXIN]: "微信调用错误",
-	[CODE.WEIXIN_LOGIN]: "微信登录错误",
-	[CODE.WEIXIN_QR_CODE]: "二维码识别出错",
-	[CODE.WEIXIN_STORAGE_SET]: "微信存储出错",
-	[CODE.LOGIN]: "登录失败",
+  [CODE.WEIXIN_QR_CODE]: "二维码识别出错",
 	[CODE.HTTP]: "网络请求失败",
 	[CODE.HTTP_PARAMS]: "内部参数错误",
-	[CODE.HTTP_API_ACCESS_TOKEN_LOGIN]: "登录失败请重开",
+  [CODE.HTTP_API_ACCESS_TOKEN_LOGIN]: "登录失败请重开",
+  [CODE.HTTP_API_ACCESS_TOKEN_VALID]: "登录验证失败",
 	[CODE.HTTP_API_USER_DETAIL]: "查询详情失败",
 	[CODE.HTTP_API_USER_UPDATE]: "更新失败",
 	[CODE.HTTP_API_TOTP_ALL]: "查询列表失败",