+++ wechat/miniprogram/constant/totp.ts
@@ -1,9 +1,9 @@
-const URL = {
-	ALL: "/miniprogram/api/v1/totp/all",
-	DETAIL: "/miniprogram/api/v1/totp/detail",
-	CREATE: "/miniprogram/api/v1/totp/create",
-	UPDATE: "/miniprogram/api/v1/totp/update",
-	DELETE: "/miniprogram/api/v1/totp/delete",
+const PATH = {
+	ALL: "/api/v1/miniprogram/totp/all",
+	DETAIL: "/api/v1/miniprogram/totp/detail",
+	CREATE: "/api/v1/miniprogram/totp/create",
+	UPDATE: "/api/v1/miniprogram/totp/edit",
+	DELETE: "/api/v1/miniprogram/totp/delete",
 };
 
-export { URL };
+export { PATH };