+++ wechat/miniprogram/types/accessToken.ts
@@ -1,9 +1,13 @@
 import type { RequestData } from "./http";
 
 export interface LoginRequest extends RequestData {
-	code: string;
+  platform: string,
+  code: string;
 }
 
 export interface LoginResponse {
-	access_token: string;
+  access_token: string;
+}
+
+export interface ValidResponse {
 }