+++ CHANGELOG.md
@@ -1,12 +1,13 @@
-## TBD-v1.6.0
+## v1.6.0
 
 ### changed
 
-- change(backend): 更改 url，将 /api/v1 更改为 /api/v1/miniprogram/
+- change(backend): 更改 url，将 /api/v1 更改为 /api/v1/miniprogram/(#75)
+- change(frontend-wechat-miniprogram): 适配后端更改(#75)
 
 ### added
 
-- feat: 支持多平台
+- chore: 支持多平台底座(#75)
 
 ## v1.5.3